Create netmiko-connecthandler-template
Template for testing ssh connectivity to a device using netmiko.

Imports ConnectHandler function from netmiko. Creates a dictionary to identify a device that assigns device. The device dictionary is passed to the ConnectHandler function and assigned to the ssh variable for later use.

The ssh variable is used to enter enable mode, send the 'hostname' command and assign it to the output variable, the ssh session is gracefully disconnected,  and the output is printed to the terminal.